highres clock

+++ game/src/render/helper/src/olcHelper.cpp
@@ -10,8 +10,8 @@ void OlcHelper::FillCircleDecal(olc::PixelGameEngine* pge, const int radius, con
   const int segments = 32;
   std::vector<olc::vf2d> points;
   for (int i = 0; i < segments; i++)
-    points.push_back(olc::vf2d(position.x + radius * cosf(i * (2 * std::numbers::pi / segments)), 
-        position.y + radius * sinf(i * (2 * std::numbers::pi / segments))));
+    points.push_back(olc::vf2d(position.x + radius * cosf(i * (2 * std::numbers::pi / (float)segments)), 
+        position.y + radius * sinf(i * (2 * std::numbers::pi / (float)segments))));
 
   std::vector<olc::vf2d> uvs;
   uvs.resize(points.size());
